minor docs examples refactoring
